apps/ai-service/requirements.txt
fastapi
uvicorn[standard]
openai
httpx
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi